Use IAM and key/secret
