+++ CloudStorageCore/src/main/java/io/github/ruschecker/transfer/TransferProgressFileInputStream.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright 2018 Emmanouil Gkatziouras
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.github.ruschecker.transfer;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+
+public final class TransferProgressFileInputStream extends FileInputStream {
+
+    private final TransferProgress transferProgress;
+
+    public TransferProgressFileInputStream(File file, TransferProgress transferProgress) throws FileNotFoundException {
+        super(file);
+        this.transferProgress = transferProgress;
+    }
+
+    @Override
+    public int read() throws IOException {
+        int b = super.read();
+        this.transferProgress.progress(new byte[]{(byte) b}, 1);
+        return b;
+    }
+
+    @Override
+    public int read(byte b[]) throws IOException {
+        int count = super.read(b);
+        this.transferProgress.progress(b, b.length);
+        return count;
+    }
+
+    @Override
+    public int read(byte b[], int off, int len) throws IOException {
+        int count = super.read(b, off, len);
+        if (off == 0) {
+            this.transferProgress.progress(b, len);
+        } else {
+            byte[] bytes = new byte[len];
+            System.arraycopy(b, off, bytes, 0, len);
+            this.transferProgress.progress(bytes, len);
+        }
+        return count;
+    }
+}
