+++ S3StorageWagon/pom.xml
@@ -2,8 +2,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <artifactId>cloud-storage-maven</artifactId>
-        <groupId>com.gkatzioura.maven.cloud</groupId>
-        <version>1.2-SNAPSHOT</version>
+        <groupId>io.github.ruschecker</groupId>
+        <version>1.2</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
@@ -34,7 +34,7 @@
             <version>${aws-s3.version}</version>
         </dependency>
         <dependency>
-            <groupId>com.gkatzioura.maven.cloud</groupId>
+            <groupId>io.github.ruschecker</groupId>
             <artifactId>cloud-storage-core</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
